requirements.txt
django==1.8.19
djangorestframework